// src/app/models/ItensLocacaoTabelas.ts
export class ItensLocacaoTabelas {
    itensLocacaoTabelaId: number;

    descricao: string;

    sedeId: number;

    tipoItemId: number;

    tipoItem: string;

    dataInclusao: Date | string;

    dataAlteracao: Date | string;

    usuarioIdInclusao: number;

    usuarioIdAlteracao: number;

    dataInativacao: Date | string | null;

    usuarioIdInativacao: number | null;

    ativo: boolean;

    nomeSede: string;

    nomeItensLocacaoTabelas: string;

}